requirements.txt
flaskflask-corsgoogle-generativeai>=0.7.0gunicornrequestsbeautifulsoup4